Includes cluster in the hostname for DNS uniqueness
This mimics what happens with the old builder, where we had
`develop--end2end.lax.elifesciences.org` and
`develop--ci.lax.elifesciences.org` available.

Before this change, the domain name would be
`develop.lax.elifesciences.org` and 2 machines will conflict with each
other.

After this is in, I would go on with simplifying deploy to only take
`stackname` and `cluster` (trying to leave the option of an additional `number` open for the future if we have multiple machines for failover/load balancing/etc).