Selecting aws-alt automatically using cluster name
In the new builder formulas `master` is the default branch for every
project, so they come up by default with the alternate configuration
(including RDS), which is undesirable.

By selecting the alternate configuration using the cluster name instead,
we can specify to use RDS only in `end2end` and `prod`, for example.

Alternate configurations that do not match a cluster name, such as
`large`, can still be selected manually using `aws_launch_instance`
instead of `deploy`.