This is a long story...
With the default behavior of SNS to SQS fanout, the messages you put
into an SNS topic will be receives by the SQS queue with a double JSON
encoding. Suppose you put into the topic:
```
{"key":"value"}
```
In the queue, you receive:
```
{
    "Type" : "Notification",
    "MessageId" : "19dce1c7-33aa-555b-a503-b1981b2544f4",
    "TopicArn" : "arn:aws:sns:us-east-1:...:bus-articles--end2end",
    "Message" : "{\"key\":\"value\"}",
    "Timestamp" : "2016-09-23T12:55:43.508Z",
    "SignatureVersion" : "1",
    "Signature" : "...",
    "SigningCertURL" : "...",
    "UnsubscribeURL" : "..."
}
```
That's a JSON document with a string property `Message` that is another
JSON-document. This is not what you want.

It turns out you can avoid this behavior with some tweaking of the
attributes of the subscription:
https://sdrbrg.se/2016/2/14/raw-sns-fanout-to-sqs/

The obscure `RawMessageDelivery` attribute is documented in an Appendix
which seemed to relate to large payloads, but in fact talks almost only
about `RawMessageDelivery` itself:
http://docs.aws.amazon.com/sns/latest/dg/large-payload-raw-message.html

Since here we use Boto 2.x (not sure if 3.x is better), we cannot set
that property. So we rewrote this monkeypatch to be able to add the
attribute:
https://gist.github.com/seizethedave/6c69c7f613664e80be11

And now we can finally get the correct message out of SQS:
```
{"key":"value"}
```

After all this is *Simple* Queue Service...