More robust algorithm to determine latest node
It may happen that the EC2 data contain a `terminated` node recently substituted by a new one which is now in a `running` (or similarly alive) state. These `terminated` nodes should be excluded from our start/stop commands.

However, if we find just a `terminated` node we should not exclude it: it's a sign we are trying to start/stop a node which is permanently dead.