Upgrade Python to 2.7.12
Python is such a hard dependency of Ubuntu and Salt that it's safer to
update it as part of the base packages instead of using a Salt state.

This is also a portable way of doing it no matter what is the basebox or
AMI being used.

Tangentially, also the virtualenvs have to be recreated after upgrading,
because the `pip` installed there complains about not being able to find
`urllib3` as part of its `requests` Python module dependency. This is
somewhat due to distro maintainers un-vendoring dependencies and taking
them out of the packages themselves; also, so much for virtual env being
an isolated Python.