Manipulating buildvars with the bootstrap users
If you think about it, this makes sense as EC2 instances have build vars right when they are created, without waiting for Salt to run. So we should be able to manipulate them at that level, without relying on the Salt update to deliver the `elife` user and access to it.

The keys of users configured in `pillar.elife` are added to both deploy and bootstrap users, so this works for anyone who was running buildvars commands (like Alfred) before.