Log Salt failures on Loggly
So that we can alert on them. Since the output can be quite lengthy, we
output to a file and only log the path to it.

We also leave the bash colors in the log file, if the file is `cat`ted
it will reproduce what you would have seen by running salt manually.