Start nodes caught when stopping
Rarely, a node takes a long time to stop and the Jenkins builds that does so times out, unlocking it. If a start task is then executed it fails. We support also the stopping state now, so that if a node is caught stopping it will be waited upon and then restarted as requested.

This happened both to me and @thewilkybarkid