DefaultTTL set to 0 seconds
If a response does not have an explicit Cache-Control header set, the current behavior is to cache it for 5 minutes, with unpleasant results in the case of the missing assets that we had during deploy.
This behavior is safer than blindly caching everything, and it makes us consider explicitly which headers to set.