Allow missing context file
We have old servers like master-server--2016-07-22 that have no context file stored on S3. Strange, we could repair them, but in general we may have the problem if the settings impose to not upload the context in the S3 bucket; at the moment 'not uploading anything to S3' is the only builder setting supported for organizations different from eLife as the bucket name is hardcoded

This is an attempt to be more conservative and still allow `bldr ssh`
and `bldr update` to work. However we may want a different direction
like making context files mandatory (hence if you have no S3 upload you
have to use the builder from a single place) and repair the missing
ones we do not have.