Attempt to remove remote_master
...and consequently the builder deploy on master-server (TBD).

We try to download the template from builder-private, render it with the formula folders (the only thing that varies) and upload it again. However, the Vagrantfile would have to do the same.

An alternative would be to write this file from within update-master.sh, but parsing and modifying a YAML file from bash didn't seem very robust. You can execute a bunch of  commands but they'll soon break horribly if a space is misplaced.