Solve bug on unsub_sqs called on a queue not subscribed to anything
The `Policy` attribute cannot be empty, so it needs to be removed by setting an empty string when there are no topics to subscribe to. Incidentally, even if a queue subscribed to topics, it will still run `unsub_sqs` at its creation when it's not subscribed to anything yet, running into the problem.