Deal with no instances (temporarily) being found
During startup it may happen that the EC2 API shows an eventually
consistent behavior and returns an empty set of EC2 instances despite
the fact that we have already seen all of them as `running` e.g.
https://alfred.elifesciences.org/blue/organizations/jenkins/test-elife-xpub/detail/test-elife-xpub/546/pipeline

Here we return an empty dictionary from `stack_all_ec2_nodes` which is
the same type as the positive result, rather than a `None`.

From the polling layer, this condition leads to wait some more until
servers appear.