Remove Vault access from masterless servers
It's not used, because we currently use Vault to provide pillars through https://github.com/elifesciences/master-server-formula/blob/master/salt/master-server/config/opt-salt-extension-modules-pillar-elife_vault.py which is only deployed to the `master-server`.

What this allows is to access Vault directly with `{{
salt['vault.read_secret'](...) }}` which we don't use, in favor of
pillars.