project file, marked unique configurations as such (#802)
* project file, marked unique configurations as such
* projects, unique config protection
* fixes deprecation warnings.
* .gitignore, adds builder-private and builder-configuration back in
* TODO, adds one.
* test_cfn.py, shifts some tests out of the class TestCfn
* .run-tests.sh, fixes bug where passing more than 2 args caused a bash warning
* checks.py, removes some very old unused checks that never came to be
* bug, fixes case where modified cached project data was being shared.
project.project_map cache results were being modified and then anything accessing them were accessing the modified version. Probably not an issue day-to-day, but definitely during testing.
The caching was put there to reduce some of the time it takes to find, parse, merge the project data.
The caching has been left in place but a deepcopy of the cached object is returned normally now.
* project file, fixed bug when 'ec2: true'. it prevents config from being merged correctly.
I suspect it had been passing tests for so long because of the project data caching bug.