Increase disk space allocated to elife-libraries--spectrum to 20gb
We tried to deploy a change to elife-spectrum recently and it ran out of disk space. We cleared out the `/ext/tmp-spectrum` folder which freed up 2GB but it filled up again in the next run.

It's evident that it needs more disk space as 5GB roughly is taken up with the cached-repositories as the elife-articles-xml is ever increasing.

This is preventing necessary changes to elife-spectrum being deployed and it is causing end2end tests to fail.